requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
et-xmlfile==1.1.0
idna==3.4
oauthlib==3.2.2
openpyxl==3.1.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
